# Proof of concept to replace `run_blocking` with something easier to use