--- a/examples/armv4t/emu.rs
+++ b/examples/armv4t/emu.rs
@@ -1,18 +1,14 @@
 use armv4t_emu::{reg, Cpu, ExampleMem, Memory, Mode};
+use gdbstub::common::Signal;
+use gdbstub::stub::run::RunTarget;
+use gdbstub::stub::SingleThreadStopReason;
+use gdbstub::target::ext::breakpoints;
+use gdbstub::target::Target;
 
 use crate::mem_sniffer::{AccessKind, MemSniffer};
 use crate::DynResult;
 
 const HLE_RETURN_ADDR: u32 = 0x12345678;
-
-#[derive(Debug, Copy, Clone, PartialEq, Eq)]
-pub enum Event {
-    DoneStep,
-    Halted,
-    Break,
-    WatchWrite(u32),
-    WatchRead(u32),
-}
 
 pub enum ExecMode {
     Step,
@@ -94,9 +90,12 @@ impl Emu {
         self.cpu.reg_set(Mode::User, reg::PC, self.start_addr);
         self.cpu.reg_set(Mode::User, reg::CPSR, 0x10);
     }
+}
 
-    /// single-step the interpreter
-    pub fn step(&mut self) -> Option<Event> {
+impl RunTarget for Emu {
+    type StopReason = SingleThreadStopReason<u32>;
+
+    fn step(&mut self) -> Result<Option<Self::StopReason>, <Self as Target>::Error> {
         let mut hit_watchpoint = None;
 
         let mut sniffer = MemSniffer::new(&mut self.mem, &self.watchpoints, |access| {
@@ -110,73 +109,42 @@ impl Emu {
             let fixup = if self.cpu.thumb_mode() { 2 } else { 4 };
             self.cpu.reg_set(Mode::User, reg::PC, pc - fixup);
 
-            return Some(match access.kind {
-                AccessKind::Read => Event::WatchRead(access.addr),
-                AccessKind::Write => Event::WatchWrite(access.addr),
-            });
+            return Ok(Some(match access.kind {
+                AccessKind::Read => SingleThreadStopReason::Watch {
+                    tid: (),
+                    kind: breakpoints::WatchKind::Read,
+                    addr: access.addr,
+                },
+                AccessKind::Write => SingleThreadStopReason::Watch {
+                    tid: (),
+                    kind: breakpoints::WatchKind::Write,
+                    addr: access.addr,
+                },
+            }));
         }
 
         if self.breakpoints.contains(&pc) {
-            return Some(Event::Break);
+            return Ok(Some(SingleThreadStopReason::SwBreak(())));
         }
 
         if pc == HLE_RETURN_ADDR {
-            return Some(Event::Halted);
+            return Ok(Some(SingleThreadStopReason::Terminated(Signal::SIGSTOP)));
         }
 
-        None
-    }
-
-    /// run the emulator in accordance with the currently set `ExecutionMode`.
-    ///
-    /// since the emulator runs in the same thread as the GDB loop, the emulator
-    /// will use the provided callback to poll the connection for incoming data
-    /// every 1024 steps.
-    pub fn run(&mut self, mut poll_incoming_data: impl FnMut() -> bool) -> RunEvent {
         match self.exec_mode {
-            ExecMode::Step => RunEvent::Event(self.step().unwrap_or(Event::DoneStep)),
-            ExecMode::Continue => {
-                let mut cycles = 0;
-                loop {
-                    if cycles % 1024 == 0 {
-                        // poll for incoming data
-                        if poll_incoming_data() {
-                            break RunEvent::IncomingData;
-                        }
-                    }
-                    cycles += 1;
-
-                    if let Some(event) = self.step() {
-                        break RunEvent::Event(event);
-                    };
-                }
-            }
-            // just continue, but with an extra PC check
+            ExecMode::Step => Ok(Some(SingleThreadStopReason::DoneStep)),
+            ExecMode::Continue => Ok(None),
             ExecMode::RangeStep(start, end) => {
-                let mut cycles = 0;
-                loop {
-                    if cycles % 1024 == 0 {
-                        // poll for incoming data
-                        if poll_incoming_data() {
-                            break RunEvent::IncomingData;
-                        }
-                    }
-                    cycles += 1;
-
-                    if let Some(event) = self.step() {
-                        break RunEvent::Event(event);
-                    };
-
-                    if !(start..end).contains(&self.cpu.reg_get(self.cpu.mode(), reg::PC)) {
-                        break RunEvent::Event(Event::DoneStep);
-                    }
+                if (start..end).contains(&pc) {
+                    Ok(None)
+                } else {
+                    Ok(Some(SingleThreadStopReason::DoneStep))
                 }
             }
         }
     }
-}
 
-pub enum RunEvent {
-    IncomingData,
-    Event(Event),
+    fn interrupt_received(&mut self) -> Result<Option<Self::StopReason>, <Self as Target>::Error> {
+        Ok(Some(SingleThreadStopReason::Signal(Signal::SIGINT)))
+    }
 }