--- a/examples/armv4t/main.rs
+++ b/examples/armv4t/main.rs
@@ -7,11 +7,10 @@ use std::net::{TcpListener, TcpStream};
 #[cfg(unix)]
 use std::os::unix::net::{UnixListener, UnixStream};
 
-use gdbstub::common::Signal;
-use gdbstub::conn::{Connection, ConnectionExt};
+use gdbstub::conn::ConnectionExt;
+use gdbstub::stub::run::RunTarget;
 use gdbstub::stub::SingleThreadStopReason;
-use gdbstub::stub::{run_blocking, DisconnectReason, GdbStub, GdbStubError};
-use gdbstub::target::Target;
+use gdbstub::stub::{DisconnectReason, GdbStubError};
 
 type DynResult<T> = Result<T, Box<dyn std::error::Error>>;
 
@@ -51,96 +50,6 @@ fn wait_for_uds(path: &str) -> DynResult<UnixStream> {
     Ok(stream)
 }
 
-enum EmuGdbEventLoop {}
-
-impl run_blocking::BlockingEventLoop for EmuGdbEventLoop {
-    type Target = emu::Emu;
-    type Connection = Box<dyn ConnectionExt<Error = std::io::Error>>;
-    type StopReason = SingleThreadStopReason<u32>;
-
-    #[allow(clippy::type_complexity)]
-    fn wait_for_stop_reason(
-        target: &mut emu::Emu,
-        conn: &mut Self::Connection,
-    ) -> Result<
-        run_blocking::Event<SingleThreadStopReason<u32>>,
-        run_blocking::WaitForStopReasonError<
-            <Self::Target as Target>::Error,
-            <Self::Connection as Connection>::Error,
-        >,
-    > {
-        // The `armv4t` example runs the emulator in the same thread as the GDB state
-        // machine loop. As such, it uses a simple poll-based model to check for
-        // interrupt events, whereby the emulator will check if there is any incoming
-        // data over the connection, and pause execution with a synthetic
-        // `RunEvent::IncomingData` event.
-        //
-        // In more complex integrations, the target will probably be running in a
-        // separate thread, and instead of using a poll-based model to check for
-        // incoming data, you'll want to use some kind of "select" based model to
-        // simultaneously wait for incoming GDB data coming over the connection, along
-        // with any target-reported stop events.
-        //
-        // The specifics of how this "select" mechanism work + how the target reports
-        // stop events will entirely depend on your project's architecture.
-        //
-        // Some ideas on how to implement this `select` mechanism:
-        //
-        // - A mpsc channel
-        // - epoll/kqueue
-        // - Running the target + stopping every so often to peek the connection
-        // - Driving `GdbStub` from various interrupt handlers
-
-        let poll_incoming_data = || {
-            // gdbstub takes ownership of the underlying connection, so the `borrow_conn`
-            // method is used to borrow the underlying connection back from the stub to
-            // check for incoming data.
-            conn.peek().map(|b| b.is_some()).unwrap_or(true)
-        };
-
-        match target.run(poll_incoming_data) {
-            emu::RunEvent::IncomingData => {
-                let byte = conn
-                    .read()
-                    .map_err(run_blocking::WaitForStopReasonError::Connection)?;
-                Ok(run_blocking::Event::IncomingData(byte))
-            }
-            emu::RunEvent::Event(event) => {
-                use gdbstub::target::ext::breakpoints::WatchKind;
-
-                // translate emulator stop reason into GDB stop reason
-                let stop_reason = match event {
-                    emu::Event::DoneStep => SingleThreadStopReason::DoneStep,
-                    emu::Event::Halted => SingleThreadStopReason::Terminated(Signal::SIGSTOP),
-                    emu::Event::Break => SingleThreadStopReason::SwBreak(()),
-                    emu::Event::WatchWrite(addr) => SingleThreadStopReason::Watch {
-                        tid: (),
-                        kind: WatchKind::Write,
-                        addr,
-                    },
-                    emu::Event::WatchRead(addr) => SingleThreadStopReason::Watch {
-                        tid: (),
-                        kind: WatchKind::Read,
-                        addr,
-                    },
-                };
-
-                Ok(run_blocking::Event::TargetStopped(stop_reason))
-            }
-        }
-    }
-
-    fn on_interrupt(
-        _target: &mut emu::Emu,
-    ) -> Result<Option<SingleThreadStopReason<u32>>, <emu::Emu as Target>::Error> {
-        // Because this emulator runs as part of the GDB stub loop, there isn't any
-        // special action that needs to be taken to interrupt the underlying target. It
-        // is implicitly paused whenever the stub isn't within the
-        // `wait_for_stop_reason` callback.
-        Ok(Some(SingleThreadStopReason::Signal(Signal::SIGINT)))
-    }
-}
-
 fn main() -> DynResult<()> {
     pretty_env_logger::init();
 
@@ -161,13 +70,15 @@ fn main() -> DynResult<()> {
         }
     };
 
-    let gdb = GdbStub::new(connection);
-
-    match gdb.run_blocking::<EmuGdbEventLoop>(&mut emu) {
+    match gdbstub::stub::run::run(connection, &mut emu) {
         Ok(disconnect_reason) => match disconnect_reason {
             DisconnectReason::Disconnect => {
                 println!("GDB client has disconnected. Running to completion...");
-                while emu.step() != Some(emu::Event::Halted) {}
+                loop {
+                    if matches!(emu.step()?, Some(SingleThreadStopReason::Terminated(_))) {
+                        break;
+                    }
+                }
             }
             DisconnectReason::TargetExited(code) => {
                 println!("Target exited with code {}!", code)