@daniel5151 2022-10-22
I've reviewed this code, and my gut feeling is that this isn't a direction I'd want to go in...

It seems you've fallen into the same trap as I did when first starting work on `gdbstub`: you've assumed that most targets operate on a "single-step" execution model, when in reality, most targets actually use the "resume + eventually stop" model. 

In fact, the design in this PR is very reminiscent of the API in `gdbstub` version 0.1, before 5278cd7 landed (prior to the release of 0.2). At the time, a step-based `gdbstub` API seemed very natural to me, given that I was integrating `gdbstub` with my own simple emulation side-projects, where everything ran on a single thread, and it was easy to interleave core execution with various GDB things (like checking for incoming interrupts, handling client requests, etc...). It was only later that I realized the limitation of this API (in the context of more advanced targets, including multi-core ones), and switched over to the current resume based API.

So yes, while we _could_ provide this kind of `run_single_step` helper API _alongside_ the current `run_blocking` API, I'm not entirely sure that's infrastructure I'd want to support (or even promote?) in `gdbstub` core, given that it only really works for the simplest of target implementations.

also, wrt. polling the connection for interrupts: that's something I've tried _very hard_ to prevent `gdbstub` from taking a stance on, and I wouldn't want to have an official API impose a "peek the connection every once in a while" model.

* * *

If I had to guess: it seems you might've over-indexed on how the armv4t example code is implemented (in `emu.rs`). Maybe I need to add more clarification docs, but that code under `emu.rs` should essentially just be a "black box" that simulates the typical API most targets have (i.e: targets that base execution around the `.run()` method, and don't directly expose a `.step()` method - rather, simulating single-step by setting an internal flag that causes `.run()` to only run for a single step when invoked)