@geigerzaehler 2022-10-23
Thanks for the reply. I see your point, in particular the one about the assumption of a `step()` method.